+++ RELEASE_NOTES.md
@@ -30,6 +30,7 @@ This version also moves the basic `KeyboardTheme` model from KeyboardKit Pro to 
 * `Keyboard.NumberPad` is a new keyboard view.
 * `KeyboardApp` has a new `keyboardExtensionBundleId` property.
 * `KeyboardApp` has a new `keyboardExtensionBundleIdWildcard` property.
+* `KeyboardSettings` & `DictationContext` can now be set up for an app.
 * `KeyboardThemeContext` is a new context type for theme-related state.
 * `KeyboardView` now applies next character probabilities to its items.
 * `KeyboardViewItem` now applies next character probability to the button modifier. 
@@ -44,6 +45,7 @@ This version also moves the basic `KeyboardTheme` model from KeyboardKit Pro to 
 
 * `Color` extensions are now moved to the `Styling` namespace.
 * `Image` extensions are now moved to the `Styling` namespace.
+* `KeyboardApp` `licenseKey` is no longer optional, but defaults to empty.
 
 ### 🗑️ Deprecations & Renamings